Refactor TickTask
