+++ src/main/kotlin/gg/skytils/skytilsmod/Skytils.kt
@@ -60,6 +60,8 @@ import gg.skytils.skytilsmod.mixins.transformers.accessors.AccessorSettingsGui
 import gg.skytils.skytilsmod.utils.*
 import gg.skytils.skytilsmod.utils.graphics.ScreenRenderer
 import gg.skytils.skytilsmod.utils.graphics.colors.CustomColor
+import gg.skytils.skytilsmod.utils.tasks.TickTask
+import gg.skytils.skytilsmod.utils.tasks.TickTaskManager
 import io.ktor.client.*
 import io.ktor.client.engine.cio.*
 import io.ktor.client.plugins.*
