+++ src/main/kotlin/gg/skytils/skytilsmod/features/impl/farming/FarmingFeatures.kt
@@ -25,11 +25,11 @@ import gg.skytils.skytilsmod.Skytils.Companion.prefix
 import gg.skytils.skytilsmod.Skytils.Companion.successPrefix
 import gg.skytils.skytilsmod.core.DataFetcher
 import gg.skytils.skytilsmod.core.SoundQueue
-import gg.skytils.skytilsmod.core.TickTask
 import gg.skytils.skytilsmod.events.impl.PacketEvent.ReceiveEvent
 import gg.skytils.skytilsmod.features.impl.handlers.MayorInfo
 import gg.skytils.skytilsmod.utils.Utils
 import gg.skytils.skytilsmod.utils.stripControlCodes
+import gg.skytils.skytilsmod.utils.tasks.TickTask
 import net.minecraft.client.gui.GuiChat
 import net.minecraft.network.play.server.S45PacketTitle
 import net.minecraft.util.ChatComponentText
