+++ src/main/kotlin/gg/skytils/skytilsmod/features/impl/farming/GardenFeatures.kt
@@ -23,8 +23,8 @@ import gg.essential.universal.utils.MCClickEventAction
 import gg.essential.universal.wrappers.message.UTextComponent
 import gg.skytils.skytilsmod.Skytils
 import gg.skytils.skytilsmod.Skytils.Companion.mc
-import gg.skytils.skytilsmod.core.TickTask
 import gg.skytils.skytilsmod.utils.*
+import gg.skytils.skytilsmod.utils.tasks.TickTask
 import net.minecraft.init.Blocks
 import net.minecraft.util.BlockPos
 import net.minecraft.util.MovingObjectPosition
