+++ src/main/kotlin/gg/skytils/skytilsmod/features/impl/misc/BrewingFeatures.kt
@@ -21,10 +21,10 @@ package gg.skytils.skytilsmod.features.impl.misc
 import gg.essential.universal.UMatrixStack
 import gg.skytils.skytilsmod.Skytils
 import gg.skytils.skytilsmod.Skytils.Companion.mc
-import gg.skytils.skytilsmod.core.TickTask
 import gg.skytils.skytilsmod.events.impl.GuiContainerEvent
 import gg.skytils.skytilsmod.events.impl.PacketEvent
 import gg.skytils.skytilsmod.utils.*
+import gg.skytils.skytilsmod.utils.tasks.TickTask
 import net.minecraft.inventory.ContainerChest
 import net.minecraft.network.play.client.C08PacketPlayerBlockPlacement
 import net.minecraft.tileentity.TileEntityBrewingStand
