+++ src/main/kotlin/gg/skytils/skytilsmod/features/impl/misc/SlayerFeatures.kt
@@ -28,7 +28,6 @@ import gg.skytils.skytilsmod.Skytils.Companion.prefix
 import gg.skytils.skytilsmod.core.GuiManager
 import gg.skytils.skytilsmod.core.GuiManager.createTitle
 import gg.skytils.skytilsmod.core.SoundQueue
-import gg.skytils.skytilsmod.core.TickTask
 import gg.skytils.skytilsmod.core.structure.GuiElement
 import gg.skytils.skytilsmod.events.impl.BlockChangeEvent
 import gg.skytils.skytilsmod.events.impl.CheckRenderEntityEvent
@@ -44,6 +43,7 @@ import gg.skytils.skytilsmod.utils.ScoreboardUtil.sidebarLines
 import gg.skytils.skytilsmod.utils.graphics.ScreenRenderer
 import gg.skytils.skytilsmod.utils.graphics.SmartFontRenderer
 import gg.skytils.skytilsmod.utils.graphics.colors.CommonColors
+import gg.skytils.skytilsmod.utils.tasks.TickTask
 import kotlinx.coroutines.*
 import net.minecraft.block.*
 import net.minecraft.client.gui.GuiChat
