+++ src/main/kotlin/gg/skytils/skytilsmod/utils/tasks/TickTask.kt
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package gg.skytils.skytilsmod.core
+package gg.skytils.skytilsmod.utils.tasks
 
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
 import net.minecraftforge.fml.common.gameevent.TickEvent
